refactor: Move complicated Paths to submodules
